create ModelController
