+++ examples/advanced/cifar10/cifar10-sim/jobs/cifar10_fedavg/cifar10_fedavg/config/config_fed_server.json
@@ -2,7 +2,7 @@
     "format_version": 2,
     "min_clients": 8,
     "num_rounds": 50,
-    "TRAIN_SPLIT_ROOT": "/tmp/cifar10_splits/cifar10_fedavg_alpha0.1_74ce0975-45f9-4982-838d-a32120ed1a9e",
+    "TRAIN_SPLIT_ROOT": "/tmp/cifar10_splits/cifar10_fedavg_alpha0.1_018566b9-8edc-4462-956f-0ad2021ddb77",
     "alpha": 0.1,
     "server": {
         "heart_beat_timeout": 600
@@ -55,8 +55,7 @@
                 "min_clients": "{min_clients}",
                 "num_rounds": "{num_rounds}",
                 "wait_time_after_min_received": 10,
-                "persistor_id": "persistor",
-                "train_timeout": 0
+                "persistor_id": "persistor"
             }
         },
         {
