+++ examples/advanced/cifar10/cifar10-sim/jobs/cifar10_scaffold/cifar10_scaffold/config/config_fed_client.json
@@ -1,6 +1,6 @@
 {
     "format_version": 2,
-    "TRAIN_SPLIT_ROOT": "/tmp/cifar10_splits/cifar10_scaffold_alpha0.1_ca1aa955-1a35-4bcf-8863-6ed24d9dd2f6",
+    "TRAIN_SPLIT_ROOT": "/tmp/cifar10_splits/cifar10_scaffold_alpha0.1_283af132-e164-4661-b634-42e74ae9e969",
     "AGGREGATION_EPOCHS": 4,
     "executors": [
         {
