+++ examples/advanced/cifar10/cifar10-sim/jobs/cifar10_scaffold/cifar10_scaffold/config/config_fed_server.json
@@ -2,7 +2,7 @@
     "format_version": 2,
     "min_clients": 8,
     "num_rounds": 50,
-    "TRAIN_SPLIT_ROOT": "/tmp/cifar10_splits/cifar10_scaffold_alpha0.1_ca1aa955-1a35-4bcf-8863-6ed24d9dd2f6",
+    "TRAIN_SPLIT_ROOT": "/tmp/cifar10_splits/cifar10_scaffold_alpha0.1_283af132-e164-4661-b634-42e74ae9e969",
     "alpha": 0.1,
     "server": {
         "heart_beat_timeout": 600
@@ -55,8 +55,7 @@
                 "min_clients": "{min_clients}",
                 "num_rounds": "{num_rounds}",
                 "wait_time_after_min_received": 10,
-                "persistor_id": "persistor",
-                "train_timeout": 0
+                "persistor_id": "persistor"
             }
         },
         {
